Change type to int8_t

+++ evalAssignment1/update_locations.h
@@ -4,9 +4,9 @@
 #include<stdlib.h>
 #include<time.h>
 
-double * generate_random_list(int32_t size, int32_t bound);
+int8_t * generate_random_list(int32_t size, int32_t bound);
 
-void update_coords(double * xs, double * ys, double * zs, double * vx, double * vy, double * vz, int32_t size);
+void update_coords(int8_t * xs, int8_t * ys, int8_t * zs, int8_t * vx, int8_t * vy, int8_t * vz, int32_t size);
 
-double sum(double * array, int32_t size);
+int8_t sum(int8_t * array, int32_t size);
